// smart-shopping/src/app/models/product.ts
export interface product{
    id: number;
    productName: String;
    productType: number;
    brand: String;
    quantityType: number;
    ratePerQuantity: number;
    stockCount:number;
    addDate: Date;
    aisle : String;
    shelf: String;
    dateOfManufacture: Date;
    dateOfExpiry: Date;
    image: String;
}